backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cfgv==3.4.0
cryptography==42.0.8
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==25.8.0
filelock==3.14.0
identify==2.5.36
nodeenv==1.8.0
platformdirs==4.2.1
pre-commit==3.7.1
psycopg==3.1.19
psycopg-binary==3.1.19
pycparser==2.22
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
virtualenv==20.26.1
